More protection against objc only routes

+++ setup/TemplateConfigurator.rb
@@ -140,6 +140,8 @@ module Pod
 
     def customise_prefix
       prefix_path = "Example/Tests/Tests-Prefix.pch"
+      return unless File.exists? prefix_path
+      
       pch = File.read prefix_path
       pch.gsub!("${INCLUDED_PREFIXES}", @prefixes.join("\n  ") )
       File.open(prefix_path, "w") { |file| file.puts pch }
